good start

+++ components/display/user_control.h
@@ -3,7 +3,7 @@
 
 #include "stdint.h"
 
-void handle_button_press(uint8_t btn_id);
+void handle_button_press(int btn_id);
 void user_button_task(void *parameter);
 
 extern uint16_t co2_threshold;
